Eclipse IDE: rename dirs not supported
Using pom only we can change the name of resources bundled to the
bundle. This seems not to be supported by the Eclipse IDE (PDE).
So, as long as we do not know how to bundle "src/main/lib" to "lib", we
rename it to "lib" in the source tree.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>